requirements.txt
transformers
torch
torchvision
onnxruntime-gpu
onnx
pillow
numpy